Move cli/deposit.py back to eth2deposit for less confusion
