+++ Makefile
@@ -4,11 +4,12 @@ PYTHON=${VENV_NAME}/bin/python3.8
 
 help:
 	@echo "clean - remove build and Python file artifacts"
-	@echo "deposit - run deposit-cli"
-	@echo "build - install basic dependencies"
-	@echo "build_test - install testing dependencies"
-	@echo "lint - check style with flake8 and mypy"
-	@echo "test - run tests"
+	# Run with venv
+	@echo "venv_deposit - run deposit cli with venv"
+	@echo "venv_build - install basic dependencies with venv"
+	@echo "venv_build_test - install testing dependencies with venv"
+	@echo "venv_lint - check style with flake8 and mypy with venv"
+	@echo "venv_test - run tests with venv"
 
 clean:
 	rm -rf venv/
@@ -25,16 +26,16 @@ $(VENV_NAME)/bin/activate: requirements.txt
 	${VENV_NAME}/bin/python -m pip install -r requirements_test.txt
 	@touch $(VENV_NAME)/bin/activate
 
-build: $(VENV_NAME)/bin/activate
+venv_build: $(VENV_NAME)/bin/activate
 
-build_test: build
+venv_build_test: venv_build
 	${VENV_NAME}/bin/python -m pip install -r requirements_test.txt
 
-test: build_test
+venv_test: venv_build_test
 	$(VENV_ACTIVATE) && python -m pytest .
 
-lint: build_test
-	$(VENV_ACTIVATE) && flake8 --config=flake8.ini ./eth2deposit ./cli ./tests && mypy --config-file mypy.ini -p eth2deposit -p tests -p cli
+venv_lint: venv_build_test
+	$(VENV_ACTIVATE) && flake8 --config=flake8.ini ./eth2deposit ./cli ./tests && mypy --config-file mypy.ini -p eth2deposit -p tests
 
-deposit: build
-	$(VENV_ACTIVATE) && python setup.py install; python ./cli/deposit.py
+venv_deposit: venv_build
+	$(VENV_ACTIVATE) && python ./eth2deposit/deposit.py
